Update README.md
The Cyclistic Bike-Share Marketing Analysis is the first case study of Google’s Data Analytics Professional Certificate program. The case study requires the analyst to follow to the steps of the data analysis process (ask, prepare, process, analyze, share, and act) for a dataset that represents a fictional company (Cyclistic Bike-Share).